Creates a new Datawrapper chart

Creates and returns the metadata of the new Datawrapper chart.

dw_create_chart(api_key = "environment", title = "", type = "")

Arguments

api_key
Optional. A Datawrapper-API-key as character string. Defaults to "environment" - tries to automatically retrieve the key that's stored in the .Reviron-file by datawrapper_auth.

title
Optional. Will set a chart's title on creation.

type
Optional. Changes the type of the chart. See the documentation for the different types.

Value

A list with the elements from the Datawrapper-API, the same as in dw_retrieve_chart_metadata()

Note

This function retrieves all metadata about the new chart that's stored by Datawrapper. If not specified, the new chart will by default be created without a title and with the type d3-lines.

Examples

if (FALSE) dw_create_chart() # uses the preset key in the .Renviron-file
if (FALSE) dw_create_chart(title = "Testtitle")

dw_create_chart(api_key = "1234ABCD") # uses the specified key

#> [1] "New chart's id: C6JzZ"

#> $status
#> [1] "ok"
#> 
#> $data
#> $data[[1]]
#> $data[[1]]$id
#> [1] "C6JzZ"
#> 
#> $data[[1]]$title
#> [1] "[ Insert title here ]"
#> 
#> $data[[1]]$theme
#> [1] "datawrapper-data"
#> 
#> $data[[1]]$createdAt
#> [1] "2019-10-31 16:22:41"
#> 
#> $data[[1]]$lastModifiedAt
#> [1] "2019-10-31 16:22:41"
#> 
#> $data[[1]]$type
#> [1] "d3-lines"
#> 
#> $data[[1]]$metadata
#> $data[[1]]$metadata$data
#> $data[[1]]$metadata$data$transpose
#> [1] FALSE
#> 
#> $data[[1]]$metadata$data$`vertical-header`
#> [1] TRUE
#> 
#> $data[[1]]$metadata$data$`horizontal-header`
#> [1] TRUE
#> 
#> 
#> $data[[1]]$metadata$visualize
#> $data[[1]]$metadata$visualize$`highlighted-series`
#> list()
#> 
#> $data[[1]]$metadata$visualize$`highlighted-values`
#> list()
#> 
#> 
#> $data[[1]]$metadata$describe
#> $data[[1]]$metadata$describe$`source-name`
#> [1] ""
#> 
#> $data[[1]]$metadata$describe$`source-url`
#> [1] ""
#> 
#> $data[[1]]$metadata$describe$`number-format`
#> [1] "-"
#> 
#> $data[[1]]$metadata$describe$`number-divisor`
#> [1] 0
#> 
#> $data[[1]]$metadata$describe$`number-append`
#> [1] ""
#> 
#> $data[[1]]$metadata$describe$`number-prepend`
#> [1] ""
#> 
#> $data[[1]]$metadata$describe$intro
#> [1] ""
#> 
#> $data[[1]]$metadata$describe$byline
#> [1] ""
#> 
#> 
#> $data[[1]]$metadata$publish
#> $data[[1]]$metadata$publish$`embed-width`
#> [1] 600
#> 
#> $data[[1]]$metadata$publish$`embed-height`
#> [1] 400
#> 
#> 
#> $data[[1]]$metadata$annotate
#> $data[[1]]$metadata$annotate$notes
#> [1] ""
#> 
#> 
#> $data[[1]]$metadata$json_error
#> NULL
#> 
#> 
#> $data[[1]]$authorId
#> NULL
#> 
#> $data[[1]]$showInGallery
#> [1] FALSE
#> 
#> $data[[1]]$language
#> [1] "en-US"
#> 
#> $data[[1]]$guestSession
#> [1] "mjqocvspmc4qntlssj8d3dkjp7"
#> 
#> $data[[1]]$lastEditStep
#> [1] 0
#> 
#> $data[[1]]$publishedAt
#> NULL
#> 
#> $data[[1]]$publicUrl
#> [1] "https://s3.eu-central-1.amazonaws.com/datawrapper-charts/static/C6JzZ/index.html"
#> 
#> $data[[1]]$publicVersion
#> [1] 0
#> 
#> $data[[1]]$organizationId
#> NULL
#> 
#> $data[[1]]$forkedFrom
#> NULL
#> 
#> $data[[1]]$externalData
#> NULL
#> 
#> $data[[1]]$forkable
#> [1] FALSE
#> 
#> $data[[1]]$isFork
#> [1] FALSE
#> 
#> $data[[1]]$inFolder
#> NULL
#> 
#> $data[[1]]$utf8
#> [1] FALSE
#> 
#> 
#>