Lists all charts

Returns all created Datawrapper charts by the current user.

dw_list_charts(
  api_key = "environment",
  userId = "",
  published = "",
  search = "",
  order = "DESC",
  orderBy = "createdAt",
  limit = ""
)

Arguments

api_key
Required. A Datawrapper-API-key as character string. Defaults to "environment" - tries to automatically retrieve the key that's stored in the .Reviron-file by datawrapper_auth.

userId
Optional. A numeric. Only return charts created by a certain user id.

published
Optional. A string. Use true or false (lowercase) to use this filter.

search
Optional. A string. Use this filter to search for a certain character string.

order
Optional. A string. DESC by default, might be changed to "ASC".

orderBy
Optional. A string. Is set to "createdAt" by default. Might be changed to: "id", "email", "name".

limit
Optional. A numeric. Defaults to 100. Number of charts to be retrieved.

Value

A tibble of all created charts. Not including all metadata, which can be retrievd for a single chart using dw_retrieve_chart_metadata.

total

Total number of created charts.

next

Returns the API-call for the next page

Author

Benedict Witzenberger, Bob Rudis

Examples

if (FALSE) dw_list_charts("aBcDE") # uses the preset key in the .Renviron-file

if (FALSE) response_list <- dw_list_charts() # save list of charts into variable