Tests the Datawrapper-API-key

Tests the key by querying the Datawrapper-API for user information.

dw_test_key(api_key = "environment")

Arguments

api_key
Optional. A Datawrapper-API-key as character string. Defaults to "environment" - tries to automatically retrieve the key that's stored in the .Reviron-file by datawrapper_auth.

Value

A list with the elements from the Datawrapper-API

status

Returns 'ok' if the API-key used was correct.

data$user$id

Returns the internal user id.

data$user$email

The users e-mail adress.

data$user$isAdmin

Specifies, if the current user is admin in his/her organization.

data$user$organization$id

Returns the organization's id or short name.

data$user$organization$name

Returns the organization's full name.

Note

This function tests the API key by retrieving information about the current user from the API. If this works, the API-key is set correctly and ready to go.

Examples

dw_test_key() # uses the preset key in the .Renviron-file

#> $status
#> [1] "ok"
#> 
#> $data
#> $data$user
#> $data$user$id
#> [1] 66969
#> 
#> $data$user$email
#> [1] "benedict.witzenberger@sueddeutsche.de"
#> 
#> $data$user$name
#> NULL
#> 
#> $data$user$website
#> NULL
#> 
#> $data$user$socialmedia
#> NULL
#> 
#> $data$user$isLoggedIn
#> [1] TRUE
#> 
#> $data$user$isGuest
#> [1] FALSE
#> 
#> $data$user$isActivated
#> [1] TRUE
#> 
#> $data$user$isAdmin
#> [1] FALSE
#> 
#> $data$user$organization
#> $data$user$organization$id
#> [1] "sueddeutsche"
#> 
#> $data$user$organization$name
#> [1] "Süddeutsche Zeitung"
#> 
#> 
#> 
#> 

dw_test_key(api_key = "1234ABCD") # uses the specified key

#> $status
#> [1] "ok"
#> 
#> $data
#> $data$user
#> $data$user$id
#> NULL
#> 
#> $data$user$email
#> [1] "guest@datawrapper.de"
#> 
#> $data$user$name
#> NULL
#> 
#> $data$user$website
#> NULL
#> 
#> $data$user$socialmedia
#> NULL
#> 
#> $data$user$isLoggedIn
#> [1] FALSE
#> 
#> $data$user$isGuest
#> [1] TRUE
#> 
#> $data$user$isActivated
#> [1] TRUE
#> 
#> $data$user$isAdmin
#> [1] FALSE
#> 
#> 
#>